✨ feat(DBDesign): Refactor DBDesign component and introduce ERD and SchemaUpdates
- Replaced the existing ERDRenderer with a new `ERD` component for better modularity.
- Added `SchemaUpdates` component to display schema migration details.
- Introduced CSS Modules for styling in both `ERD` and `SchemaUpdates` components.
- Created mock data for migrations and review comments to enhance functionality.
- Implemented a new `MigrationsViewer` to handle migration documentation and comments.

This refactor improves the structure and maintainability of the DBDesign section in the SessionDetailPage.
