+++ frontend/apps/app/components/SessionDetailPage/components/Output/components/DBDesign/components/SchemaUpdates/MigrationsViewer/useMigrationsViewer/commentExtension.ts
