+++ frontend/apps/app/components/SessionDetailPage/components/Output/components/DBDesign/components/SchemaUpdates/index.ts
@@ -1,0 +1,1 @@
+export * from './SchemaUpdates'
