+++ frontend/apps/app/components/SessionDetailPage/components/Output/components/DBDesign/components/SchemaUpdates/mock.ts
